refac: price range wrapper 분리
